% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{blast}
\alias{blast}
\title{blast nucleotide}
\usage{
blast(type, query, ...)
}
\arguments{
\item{type}{blast type, available programs are blastn, blastp, blastx,
psiblast, tblastn, tblastx}

\item{query}{query sequences}
}
\description{
blast nucleotide
}
