% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCuttedPrimer.R
\name{getCuttedPrimer}
\alias{getCuttedPrimer}
\title{getCuttedPrimer}
\usage{
getCuttedPrimer(primer, len = 11, direction = "forward")
}
\arguments{
\item{primer}{input primer}

\item{len}{cutted length of primer}

\item{direction}{forward/reverse}
}
\value{
cutted primer (string)
}
\description{
Cut primers by specific length
}
\examples{
fprimer <-  'AGATATTGGAAC[A|T]TTATATTTTATTT[A|T]TGG'
getCuttedPrimer(fprimer, len=11, direction='forward')

}
