% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepSeqByPrimer.R
\name{grepSeqByPrimer}
\alias{grepSeqByPrimer}
\title{Grep core sequences by specific primer}
\usage{
grepSeqByPrimer(seq, fprimer, rprimer)
}
\arguments{
\item{seq}{a sequence}

\item{fprimer}{forward primer}

\item{rprimer}{reverse primer}
}
\value{
a sequence with primers
}
\description{

}
