% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeblastdb.R
\name{makeblastdb}
\alias{makeblastdb}
\title{make blastdb}
\usage{
makeblastdb(input, dbtype, parse_seqids = TRUE, hash_index = FALSE,
  ...)
}
\arguments{
\item{input}{input file}

\item{dbtype}{database type, "nucl": nucleotide}

\item{parse_seqids}{Option to parse seqid for FASTA input if set,
for all other input types seqids are parsed automatically,
default is TRUE}

\item{hash_index}{Create index of sequence hash values, default is FALSE}

\item{...}{makeblastdb arguments, available arguments are:
'gi_mask','gi_mask_name','h','help',
'input_type','logfile','mask_data','mask_desc',
'mask_id','max_file_sz','out','taxid','taxid_map',
'title','version'}

\item{input_type}{Type of the data specified in input_file. Available types are
`asn1_bin', `asn1_txt', `blastdb', `fasta'. Default is 'fasta'}
}
\value{
None
}
\description{
make blastdb
}
