% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothur.R
\name{mothur}
\alias{mothur}
\title{mothur nucleotide}
\usage{
mothur(binPath, subfunc, ...)
}
\arguments{
\item{binPath}{mothur binary directory}

\item{subfunc}{query sequences}
}
\description{
mothur nucleotide
}
\examples{
\dontrun{
mothurDir <- '~/Downloads/mothur/'
fastaFile <-  './1060501B_cut3_without_N.fas'

phylipFile <-  paste(strsplit(fastaFile, '.fas')[1], '.phylip.dist', sep='')
listFile <- paste(strsplit(fastaFile, '.fas')[1], '.phylip.fn.list', sep='')

mothur(mothurDir, subfunc='dist.seqs',
       fasta=fastaFile, calc='onegap', cutoff=0.03, countends=F, output='lt')
mothur(mothurDir, subfunc='cluster',
       phylip=phylipFile,
       method='furthest',cutoff=0.03)
mothur(mothurDir, subfunc='bin.seqs',
       list=listFile, fasta=fastaFile)
}
}
