% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputSeq.R
\name{outputSeq}
\alias{outputSeq}
\title{Clean up sequence with primers}
\usage{
outputSeq(fastaFile, outputFasta, fprimer, rprimer, duplinum, srange,
  sort = 4)
}
\arguments{
\item{fastaFile}{a fasta file path}

\item{outputFasta}{output fasta file path}

\item{fprimer}{forward primer}

\item{rprimer}{reverse primer}

\item{duplinum}{only select the sequences which duplicate
number is greater and equal than `duplinum`}

\item{srange}{minimum and maximum values of prunned sequence length, ex:
c(20, 150) means the minimum number is 20 and maximum number is 150}

\item{sort}{sort output table by id (1), pruned sequence (2),
duplicates (3) or pruned sequence length (4)}
}
\value{
data.table format sequence set
}
\description{
Clean up sequence with primers
}
