% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.myfasta.R
\name{read.myfasta}
\alias{read.myfasta}
\title{Import fasta file and convert it into data.table format}
\usage{
read.myfasta(fastaFile)
}
\arguments{
\item{fastaFile}{fasta file path}
}
\value{
a data.table
}
\description{
Import fasta file and convert it into data.table format
}
\examples{
# not run
read.myfasta('/path/to/example.fasta')
}
