% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputSeq.R
\name{outputSeq}
\alias{outputSeq}
\title{Clean up sequence with primers}
\usage{
outputSeq(fastaFile, outputFasta, fprimer, rprimer, duplinum, fill = TRUE)
}
\arguments{
\item{fastaFile}{a fasta file path, ex: /path/to/example.fasta}

\item{outputFasta}{output fasta file path, ex: /path/to/export.fasta}

\item{fprimer}{forward primer}

\item{rprimer}{reverse primer}

\item{duplinum}{only select the sequences which duplicate 
number is greater and equal than `duplinum`}

\item{fill}{fill the gap bp with N}
}
\value{
data.table format sequence set
}
\description{
Clean up sequence with primers
}
